Refactor UpgradeCTA + Modal for multiple usage in Boost
